third_party/blink/web_tests/external/wpt/editing/run/inserttext_1001-2000-expected.txt

This is a testharness.js-based test.
Found 17 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["inserttext"," "]] "<b>a</b><i>@</i><u>b</u>{}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>a</b><i>@</i><u>b</u> " but got "<b>a</b><i>@</i><u>b&nbsp;</u>"
[FAIL] [["inserttext"," "]] "a@b<b>[]c</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"mailto:a@b\\">a@b</a><b> c</b>" but got "a@b&nbsp;<b>c</b>"
[FAIL] [["inserttext","\\t"]] "http://a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://a\\">http://a</a>\\t" but got "http://a<span style=\\"white-space:pre\\">\\t</span>"
[FAIL] [["inserttext","\\f"]] "http://a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://a\\">http://a</a>\\f" but got "http://a\\f"
[FAIL] [["inserttext","   "]] "foo[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp;&nbsp;" but got "foo&nbsp; &nbsp;"
[FAIL] [["inserttext","a"]] "<p>foo</p>{}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p>a" but got "<p>fooa</p>"
[FAIL] [["inserttext","a"]] "{}<p>foo" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "a<p>foo</p>" but got "<p>afoo</p>"
[FAIL] [["inserttext","a"]] "<p>foo</p>{}<p>bar</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p>a<p>bar</p>" but got "<p>foo</p><p>abar</p>"
[FAIL] [["inserttext","a"]] "<b>foo</b>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo</b>abar" but got "<b>fooa</b>bar"
[FAIL] [["inserttext","a"]] "foo<b>{}</b>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>a</b>bar" but got "fooa<b></b>bar"
[FAIL] [["inserttext","a"]] "<a>foo</a>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a>foo</a>abar" but got "<a>fooa</a>bar"
[FAIL] [["inserttext","a"]] "<a href=/>foo[]</a>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"/\\">fooa</a>bar" but got "<a href=\\"/\\">foo</a>abar"
[FAIL] [["defaultparagraphseparator","div"],["inserttext","a"]] "<p>fo[o<p>b]ar" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","true"],["inserttext","a"]] "<p>[foo<span style=color:#aBcDeF>bar]</span>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["inserttext","a"]] "<p>foo<span style=color:#aBcDeF>[bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"color:rgb(171, 205, 239)\\">a</span></p>" but got "<p>foo<font color=\\"#abcdef\\">a</font></p>"
[FAIL] [["inserttext","a"]] "<p>foo<span style=color:#aBcDeF>{bar</span>baz}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"color:rgb(171, 205, 239)\\">a</span></p>" but got "<p>foo<font color=\\"#abcdef\\">a</font></p>"
[FAIL] [["stylewithcss","false"],["inserttext","a"]] "<p>foo<span style=color:#aBcDeF>[bar</span><span style=color:#fEdCbA>baz]</span>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"color:rgb(171, 205, 239)\\">a</span>quz</p>" but got "<p>foo<font color=\\"#abcdef\\">a</font>quz</p>"
Harness: the test ran to completion.
